chore: format, regenerate docs, fix lint nits
- prettier --write (all source files already formatted; docs/hardware.md
  and README hardware tables updated by docs:hardware regeneration)
- docs:api regenerated (API docs reflect recent status/getEngineVersion
  additions; parseD1Status + D1_STATUS_REQUEST removed as expected)
- docs:hardware regenerated hardware tables (4 files updated)
- fix: remove unnecessary type assertions in protocol.test.ts and
  duo-tape-bridge.ts (bm.data is already Uint8Array; d1-core module
  type is assignable to D1CoreModule)
- fix: use Object.entries() in fromUSBDevice() to eliminate
  noUncheckedIndexedAccess-required ! assertions on Record
- fix: len= template literal uses .toString() (restrict-template-expressions)
- chore: eslint-disable for unicorn/no-useless-undefined on required
  `T | undefined` parameter in duo-tape-bridge.test.ts
- chore: eslint-disable for no-deprecated re-export of DeviceSupport
  (retained during alias transition)
- chore: remove unused no-non-null-assertion disable in printer.test.ts

+++ packages/core/src/__tests__/protocol.test.ts
@@ -548,13 +548,13 @@ describe('encodeLabel', () => {
       // Fill the whole buffer with 0xff, then mask the trailing bits
       // in each row's last byte (per LabelBitmap invariant: bits past
       // `widthPx` stay zero).
-      (bm.data as Uint8Array).fill(0xff);
+      bm.data.fill(0xff);
       const trailingBits = labelWidthDots % 8;
       if (trailingBits !== 0) {
         const mask = (0xff << (8 - trailingBits)) & 0xff;
         for (let y = 0; y < bm.heightPx; y++) {
           const last = y * stride + (stride - 1);
-          (bm.data as Uint8Array)[last] = (bm.data[last] ?? 0) & mask;
+          bm.data[last] = (bm.data[last] ?? 0) & mask;
         }
       }
